+++ Core/LAMBDA/viz_functions/image_based/viz_schism_fim_processing/buildspec.yml
@@ -16,6 +16,3 @@ phases:
       - echo Build completed on `date`
       - echo Pushing the Docker image...
       - docker push $AWS_ACCOUNT_ID.dkr.ecr.$AWS_DEFAULT_REGION.amazonaws.com/$IMAGE_REPO_NAME:$IMAGE_TAG
-      - echo Updating lambda
-      - npm install -g serverless
-      - sls deploy
