+++ Core/LAMBDA/viz_functions/image_based/viz_schism_fim_processing/requirements.txt
@@ -1,16 +1,16 @@
 Fiona==1.8.22
+GeoAlchemy2==0.12.3
+SQLAlchemy==1.4.46
+Shapely==1.8.5.post1
+boto3==1.24.59
+botocore==1.27.59
+geopandas==0.12.2
 h5netcdf==1.1.0
-netcdf4==1.5.7
+netcdf4==1.5.8
 numpy==1.22
+psycopg2-binary==2.9.9
 rasterio==1.3.4
 rioxarray==0.13.1
 s3fs==2022.11.0
 scipy==1.9.3
-Shapely==1.8.5.post1
 xarray==2022.12.0
-geopandas==0.12.2
-SQLAlchemy==1.4.46
-GeoAlchemy2==0.12.3
-psycopg2-binary==2.9.3
-botocore==1.27.59
-boto3==1.24.59
