+++ Core/LAMBDA/viz_functions/viz_db_postprocess_sql/admin/ingest_finish.sql
@@ -1,4 +1,4 @@
-CREATE INDEX {index_name} ON {target_table} {index_columns};
+CREATE INDEX IF NOT EXISTS {index_name} ON {target_table} {index_columns};
 
 ALTER TABLE {target_table}
 ADD COLUMN reference_time TEXT DEFAULT '1900-01-01 00:00:00 UTC';
