+++ Core/StepFunctions/main.tf
@@ -18,7 +18,11 @@ variable "python_preprocessing_10GB_arn" {
   type        = string
 }
 
-variable "schism_fim_processing_arn" {
+variable "schism_fim_job_definition_arn" {
+  type        = string
+}
+
+variable "schism_fim_job_queue_arn" {
   type        = string
 }
 
@@ -141,9 +145,10 @@ resource "aws_sfn_state_machine" "schism_fim_processing_step_function" {
     role_arn = var.viz_lambda_role
 
     definition = templatefile("${path.module}/schism_fim_processing.json.tftpl", {
-        schism_fim_processing_arn = var.schism_fim_processing_arn
+        db_postprocess_sql_arn = var.db_postprocess_sql_arn
+        schism_fim_job_definition_arn  = var.schism_fim_job_definition_arn
+        schism_fim_job_queue_arn  = var.schism_fim_job_queue_arn
         optimize_rasters_arn      = var.optimize_rasters_arn
-        update_egis_data_arn  = var.update_egis_data_arn
     })
 }
